#I319:  Refactoring code
