+++ web-console-v2/src/data/chartsComponents/processing.ts
@@ -2,7 +2,7 @@ import * as _ from 'lodash';
 import dayjs from 'dayjs';
 import defaultConf from './common';
 import prettyMilliseconds from 'pretty-ms';
-import endpoints from 'data/apiEndpoints';
+import endpoints from 'constants/Endpoints';
 
 const dateFormat = 'YYYY-MM-DDT00:00:00+05:30'
