+++ web-console-v2/src/pages/DatasetListV1/createEvents.tsx
@@ -15,7 +15,6 @@ import { readJsonFileContents } from 'services/utils';
 import AnimateButton from 'components/@extended/AnimateButton';
 import RejectionFiles from 'components/third-party/dropzone/RejectionFiles';
 import Loader from 'components/Loader';
-import { DatasetStatus } from 'types/datasets';
 import { useAlert } from 'contexts/AlertContextProvider';
 
 const DatasetCreateEvents = () => {
