+++ web-console-v2/src/pages/alertManager/components/NotificationComponent.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Box, Button, Grid, Stack, Typography } from "@mui/material";
 import MUIForm from "components/form";
-import { StandardWidthButton } from "components/Styled/Buttons";
 import { useEffect, useRef, useState } from "react";
 import { fetchChannels } from "services/notificationChannels";
 import interactIds from "data/telemetry/interact.json";
