+++ web-console-v2/src/router/index.tsx
@@ -14,8 +14,8 @@ import SelectConnectorPage from 'pages/SelectConnector/SelectConnector';
 import ManageConnectorsPage from 'pages/ConnectorManagement/Manage/Manage';
 import Dashboard from 'pages/Dashboard/Dashboard';
 import IndividualMetricDashboards from 'pages/Dashboard/IndividualDashboardPage/IndividualDashboardPage';
-import DatasetCreateEvents from 'pages/dashboardV1/createEvents';
-import ClusterHealth from 'pages/dashboardV1/datasets';
+import DatasetCreateEvents from 'pages/DatasetListV1/createEvents';
+import ClusterHealth from 'pages/DatasetListV1/datasets';
 import StepperPage from 'pages/StepsPages/StepperPage';
 import AlertRules from 'pages/alertManager/views/AlertRules';
 import SystemAlerts from 'pages/alertManager/views/SystemRules';
